[IMP] mail: enhance activity management and systray interface
This commit makes the following improvements for the activities.

- Added a rainbow_man effect for empty systray opening.
- Improve calendar view systray now only 2 upcoming meetings and 1 allday
  meetings will be shown to reduce the noise in the systray, with info of
  remaining additional meeting at below them.
- Added support to reuse the done activity. Added the reuse button in the
  message actions. On clicking that activity schedule wizard will open with
  prefilled configurations of uthe sed activity. Added the mail_activity_id
  field on the message, as earlier the activity and message were never connected
  so this feature will not work for the older activities.

Task-5344015

+++ addons/mail/static/src/core/web/activity_model_patch.js
@@ -65,8 +65,10 @@ patch(Activity.prototype, {
         });
         return action;
     },
-    remove({ broadcast = true } = {}) {
-        this.delete();
+    remove(deleteActivity = true, { broadcast = true } = {}) {
+        if (deleteActivity) {
+            this.delete();
+        }
         if (broadcast) {
             this.activityBroadcastChannel?.postMessage({
                 type: "DELETE",
